Make environment separately, also move constraints to an own file.
The reason why we want to make the environment only once is to create
the constants for BAD/UNR/Bottom only once. Ideal would be to give
them some constant, wired in, unique. Anyways, the current solution
will change because:

Contract translation is on the _unlifted_ program, but the environment
is built with the _lifed_ program...