Marshall some builtins brutally, and add new flag --db-anns
Strangely, on my Arch Linux' 7.4.1 ghc the [] and : constructors are
found, but Koen reports from an unknown OS with 7.4.1 and I confirmed
this with May 2012 Haskell Platform windows that it doesn't work, so
Lists, bools and the binary tuple are brutally added.

The new flag `--db-anns' will report what ANN pragmas are found and
how they are interconnected.