Remove AEA in favour of DCA, while correctly making CF for nullary constructors
Dimitrios' observation:

[Dan's crazy Axiom (DCA)] :
       min(K xs) /\ not (cf(K xs)) =>  BigOr (min(x_i) /\ not (cf(x_i))

[Another Existing Axiom (AEA) ]:
      min(K xs) /\  (cf xs) => cf(K xs)

Claim:
     DCA => AEA

Let's do what theorem provers do, and see if DCA /\ ~AEA is
 contradictory. What is ~AEA? Well:

       AEA = min(K xs) /\ BigAnd cf(x_i) => cf(K xs)

It's on the form A -> B, and ~(A -> B) = ~(~A \/ B) = A /\ ~B, thus:

      ~AEA = min(K xs) /\ BigAnd cf(x_i) /\ ~cf(K xs)

Now, min(K xs) /\ ~cf(K xs) give us

  (*) BigOr (min(K_i) /\ ~cf(x_i))

Directly from DCA, and we still have the BigAnd cf(x_i) around, so use cf => min:

      /\ BigAnd cf(x_i) /\ BigAnd min(x_i)

But, equation (*) above says that one of these i is

 (**)  ~cf(x_i)

So we have a contradiction. but with one important catch! Consider a
nullary constructor K. We want to say that min(K) => cf(K), and this
before came (vacuously) from AEA, but now we have to add it ourself.

It can be seen it in the derivation above as well, marked (**), that I
pick an i, but if the constructor is nullary, there is no such i to pick.