Add support for paradox, and racing paradox against other provers
Since we can now get finite countersatisfiability, it may be worthwile
to test paradox. This patch adds paradox, but also the ability to run
paradox in parallel with another theorem prover, and terminate the
other when one of them finishes. You use this by using the upper case
letter corresponding to the theorem provers. So we have:

  Flags for --provers (short -p):

        -e   eprover        -E   race paradox against eprover
        -f   eproof         -F   race paradox against eproof
        -x   equinox        -X   race paradox against equinox
        -s   spass          -S   race paradox against spass
        -z   z3             -Z   race paradox against z3
        -v   vampire        -V   race paradox against vampire
        -p   paradox        -P   race paradox against paradox

Experimental results are not very encouraging.