First steps on auto mode
Apparently the unfoldings are not added for things not used enough (?).
If I comment out the signature in Integers, there are no unfoldings
at all generated (!). So we need to carry around a little map of
bindings. However, imports from other modules have seemed to work properly
before, and for example the HipSpec.Prelude imports work correctly.