Update signature map translation
Now supports built in lists and tuples, had to introduce them to the
scope by hand (?!).

As before, newtypes are considered abstract.

The quickspec term -> simple translation is still incorrect, but it works
with Alt-Ergo because the type arguments are not written out and it
does its own type inference.